backend/pyproject.toml
[project]
name = "mvp-backend"
version = "0.1.0"
description = "FastAPI backend service for the MVP project."
authors = [{ name = "Santosh" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.32.0,<0.33.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "pyjwt[crypto]>=2.9.0,<3.0.0",
    "httpx>=0.27.2,<0.28.0",
    "email-validator>=2.2.0,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "pandas>=2.2.0,<3.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "xlrd>=2.0.1,<3.0.0",
    "celery[redis]>=5.4.0,<6.0.0",
    "redis>=5.0.0,<6.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "pdfplumber>=0.11.0,<1.0.0",
    "PyMuPDF>=1.24.7,<1.25.0",
    "google-api-python-client>=2.146.0,<3.0.0",
    "google-auth>=2.35.0,<3.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "google-cloud-pubsub>=2.24.0,<3.0.0",
    "scikit-learn>=1.5.0,<2.0.0",
    "numpy>=1.26.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2,<9.0.0",
    "ruff>=0.6.9,<0.7.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = "app"
python = "3.11"

